feat: setUser if there is an active session so that it doesnt lose it on reload

+++ context/UserContext.js
@@ -1,10 +1,11 @@
-import { createContext, useState } from "react";
+import { createContext, useEffect, useState } from "react";
 import { account } from "../lib/appwrite";
 import { ID } from "react-native-appwrite";
 export const UserContext = createContext();
 
 export function UserProvider({ children }) {
   const [user, setUser] = useState(null);
+  const [authChecked, setAuthChecked] = useState(false);
   async function login(email, password) {
     try {
       await account.createEmailPasswordSession(email, password);
@@ -23,9 +24,21 @@ export function UserProvider({ children }) {
     }
   }
   async function logout() {
+    if (user === null) throw Error("cant logout non existant user");
     await account.deleteSession("current");
     setUser(null);
   }
+  async function getInitialUserValue() {
+    try {
+      const response = await account.get();
+      setUser(response);
+    } catch (error) {
+      setUser(null);
+    } finally {
+      setAuthChecked(true);
+    }
+  }
+  useEffect(getInitialUserValue, []);
   return (
     <UserContext.Provider value={{ user, login, register, logout }}>
       {children}
